Task/dr 332 lambda deploy infra (#7)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

The following AWS Lambdas are spun up via terraform and provisioned via
a CI/CD pipeline. The AWS Lambda is functionals is able to interact with
the associated existing Dynamodb tables.

- Organisations 
- Organisation_Affliation
- Locations
- Healthcare_Services 
- Questionnaire 
- Questionnaire Response

There will be additional work required later to connect the API gateways
& microservices with the Lambdas, code is submitted with this in
consideration.

## Context

This change is required in order to deploy AWS Lambda in order to make
contact and interact with the existing Dynamo DB tables.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [x] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [x] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.