Task/dr 350 deploy crud routing (#11)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

With the Lambda functions live for Capacity Management, it is necessary
to create some provisional application code for their internals.
The basic requirement for each manager Lambda is to be able to Create,
Read, Update and Delete given items in their underlying table.
The following four functions are therefore necessary in the Lambda which
perform these basic REST operations:

- A function to GET (read) HealthcareService 
- A function to POST (create) HealthcareService
- A function to PUT (update) HealthcareService 
- A function to DELETE (delete) HealthcareService



## Context

The basic requirement for each manager Lambda is to be able to Create,
Read, Update and Delete given items in their underlying table.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [x] I have added tests to cover my changes
- [x] I have updated the documentation accordingly
- [x] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.